refactor: extract project-knowledge.js from project.js (PR-09)
Move knowledge file management handlers (list, read, save, delete,
promote, depromote) into a dedicated module. project.js reduced from
3,428 to 3,278 lines.
