+++ lib/project.js
@@ -21,6 +21,7 @@ var { attachLoop } = require("./project-loop");
 var { attachFileWatch } = require("./project-file-watch");
 var { attachHTTP } = require("./project-http");
 var { attachImage } = require("./project-image");
+var { attachKnowledge } = require("./project-knowledge");
 
 // --- Context Sources persistence ---
 var _ctxSrcConfig = require("./config");
@@ -335,6 +336,15 @@ function createProjectContext(opts) {
       }
     }
   }
+
+  // --- Knowledge engine (delegated to project-knowledge.js) ---
+  var _knowledge = attachKnowledge({
+    cwd: cwd,
+    isMate: isMate,
+    sendTo: sendTo,
+    matesModule: matesModule,
+    getProjectOwnerId: function () { return projectOwnerId; },
+  });
 
   // --- File/directory watcher engine (delegated to project-file-watch.js) ---
   var _fileWatch = attachFileWatch({
@@ -856,168 +866,8 @@ function createProjectContext(opts) {
       return;
     }
 
-    // --- Knowledge file management ---
-    if (msg.type === "knowledge_list") {
-      var knowledgeDir = path.join(cwd, "knowledge");
-      var files = [];
-      try {
-        var entries = fs.readdirSync(knowledgeDir);
-        for (var ki = 0; ki < entries.length; ki++) {
-          if (entries[ki] === "session-digests.jsonl") continue;
-          if (entries[ki] === "sticky-notes.md") continue;
-          if (entries[ki] === "memory-summary.md") continue;
-          if (entries[ki].endsWith(".md") || entries[ki].endsWith(".jsonl")) {
-            var stat = fs.statSync(path.join(knowledgeDir, entries[ki]));
-            files.push({ name: entries[ki], size: stat.size, mtime: stat.mtimeMs, common: false });
-          }
-        }
-      } catch (e) { /* dir may not exist */ }
-      files.sort(function (a, b) { return b.mtime - a.mtime; });
-
-      // For mate projects, check which files are promoted and include common files from other mates
-      if (isMate) {
-        var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-        var thisMateId = path.basename(cwd);
-        // Tag promoted files
-        for (var pi = 0; pi < files.length; pi++) {
-          files[pi].promoted = matesModule.isPromoted(mateCtx, thisMateId, files[pi].name);
-        }
-        // Get common files from other mates
-        var commonFiles = matesModule.getCommonKnowledgeForMate(mateCtx, thisMateId);
-        // Filter out entries that belong to THIS mate (those are already in the list as promoted)
-        for (var ci = 0; ci < commonFiles.length; ci++) {
-          if (commonFiles[ci].ownMateId !== thisMateId) {
-            files.push(commonFiles[ci]);
-          }
-        }
-      }
-
-      sendTo(ws, { type: "knowledge_list", files: files });
-      return;
-    }
-
-    if (msg.type === "knowledge_read") {
-      if (!msg.name) return;
-      var safeName = path.basename(msg.name);
-      var filePath;
-      if (msg.common && msg.ownMateId && isMate) {
-        // Reading a common file from another mate
-        var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-        try {
-          var content = matesModule.readCommonKnowledgeFile(mateCtx, msg.ownMateId, safeName);
-          sendTo(ws, { type: "knowledge_content", name: safeName, content: content, common: true, ownMateId: msg.ownMateId });
-        } catch (e) {
-          sendTo(ws, { type: "knowledge_content", name: safeName, content: "", error: "File not found", common: true });
-        }
-      } else {
-        filePath = path.join(cwd, "knowledge", safeName);
-        try {
-          var content = fs.readFileSync(filePath, "utf8");
-          sendTo(ws, { type: "knowledge_content", name: safeName, content: content });
-        } catch (e) {
-          sendTo(ws, { type: "knowledge_content", name: safeName, content: "", error: "File not found" });
-        }
-      }
-      return;
-    }
-
-    if (msg.type === "knowledge_save") {
-      if (!msg.name || typeof msg.content !== "string") return;
-      var safeName = path.basename(msg.name);
-      if (!safeName.endsWith(".md") && !safeName.endsWith(".jsonl")) safeName += ".md";
-      var knowledgeDir = path.join(cwd, "knowledge");
-      fs.mkdirSync(knowledgeDir, { recursive: true });
-      fs.writeFileSync(path.join(knowledgeDir, safeName), msg.content);
-      // Return updated list
-      var files = [];
-      try {
-        var entries = fs.readdirSync(knowledgeDir);
-        for (var ki = 0; ki < entries.length; ki++) {
-          if (entries[ki].endsWith(".md") || entries[ki].endsWith(".jsonl")) {
-            var stat = fs.statSync(path.join(knowledgeDir, entries[ki]));
-            files.push({ name: entries[ki], size: stat.size, mtime: stat.mtimeMs });
-          }
-        }
-      } catch (e) {}
-      files.sort(function (a, b) { return b.mtime - a.mtime; });
-      // Tag files for mate projects
-      if (isMate) {
-        var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-        var thisMateId = path.basename(cwd);
-        for (var pi = 0; pi < files.length; pi++) {
-          files[pi].common = false;
-          files[pi].promoted = matesModule.isPromoted(mateCtx, thisMateId, files[pi].name);
-        }
-        var commonFiles = matesModule.getCommonKnowledgeForMate(mateCtx, thisMateId);
-        for (var ci = 0; ci < commonFiles.length; ci++) {
-          if (commonFiles[ci].ownMateId !== thisMateId) files.push(commonFiles[ci]);
-        }
-      }
-      sendTo(ws, { type: "knowledge_saved", name: safeName });
-      sendTo(ws, { type: "knowledge_list", files: files });
-      return;
-    }
-
-    if (msg.type === "knowledge_delete") {
-      if (!msg.name) return;
-      var safeName = path.basename(msg.name);
-      var filePath = path.join(cwd, "knowledge", safeName);
-      try { fs.unlinkSync(filePath); } catch (e) {}
-      // Return updated list
-      var knowledgeDir = path.join(cwd, "knowledge");
-      var files = [];
-      try {
-        var entries = fs.readdirSync(knowledgeDir);
-        for (var ki = 0; ki < entries.length; ki++) {
-          if (entries[ki].endsWith(".md") || entries[ki].endsWith(".jsonl")) {
-            var stat = fs.statSync(path.join(knowledgeDir, entries[ki]));
-            files.push({ name: entries[ki], size: stat.size, mtime: stat.mtimeMs });
-          }
-        }
-      } catch (e) {}
-      files.sort(function (a, b) { return b.mtime - a.mtime; });
-      // Tag files for mate projects
-      if (isMate) {
-        var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-        var thisMateId = path.basename(cwd);
-        for (var pi = 0; pi < files.length; pi++) {
-          files[pi].common = false;
-          files[pi].promoted = matesModule.isPromoted(mateCtx, thisMateId, files[pi].name);
-        }
-        var commonFiles = matesModule.getCommonKnowledgeForMate(mateCtx, thisMateId);
-        for (var ci = 0; ci < commonFiles.length; ci++) {
-          if (commonFiles[ci].ownMateId !== thisMateId) files.push(commonFiles[ci]);
-        }
-      }
-      sendTo(ws, { type: "knowledge_deleted", name: safeName });
-      sendTo(ws, { type: "knowledge_list", files: files });
-      return;
-    }
-
-    if (msg.type === "knowledge_promote") {
-      if (!isMate || !msg.name) return;
-      var safeName = path.basename(msg.name);
-      var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-      var thisMateId = path.basename(cwd);
-      var mate = matesModule.getMate(mateCtx, thisMateId);
-      var mateName = (mate && mate.name) || null;
-      matesModule.promoteKnowledge(mateCtx, thisMateId, mateName, safeName);
-      sendTo(ws, { type: "knowledge_promoted", name: safeName });
-      // Re-send updated list (reuse knowledge_list logic)
-      handleMessage(ws, { type: "knowledge_list" });
-      return;
-    }
-
-    if (msg.type === "knowledge_depromote") {
-      if (!isMate || !msg.name) return;
-      var safeName = path.basename(msg.name);
-      var mateCtx = matesModule.buildMateCtx(projectOwnerId);
-      var thisMateId = path.basename(cwd);
-      matesModule.depromoteKnowledge(mateCtx, thisMateId, safeName);
-      sendTo(ws, { type: "knowledge_depromoted", name: safeName });
-      handleMessage(ws, { type: "knowledge_list" });
-      return;
-    }
+    // --- Knowledge file management (delegated to project-knowledge.js) ---
+    if (_knowledge.handleKnowledgeMessage(ws, msg)) return;
 
     // --- Memory (session digests) management (delegated to project-memory.js) ---
     if (msg.type === "memory_list") { _memory.handleMemoryList(ws); return; }
